requirement.txt
accelerate==0.32.1
scipy==1.14.0
diffusers==0.29.2
transformers==4.42.4
typer==0.12.3
tqdm==4.66.4
natsort==8.4.0
open_clip_torch==2.26.1
wandb==0.17.4
peft==0.11.0
safetensors==0.4.3
hpsv2==1.2.0